Add files via upload
Our project, HirePeople, is a web application that connects job seekers with potential employers. Using React.js for the frontend, Spring Boot for the backend, and MongoDB Atlas for the database, HirePeople offers a user-friendly interface for job seekers to search and apply for jobs based on different criteria, such as location, job type, and industry. Employers can also create job listings, manage applications, and search for candidates based on their skills and experience. With HirePeople, we aim to streamline the job search process and make it easier for job seekers and employers to find the right match.